- Adds Controllers to Provide CRUD Operations - Adds Providers to Alter IO Data according to the Model's constraints - Changes Models - Issue #6

+++ Database/Repositories/PrescriptionRepository.cs
@@ -46,7 +46,11 @@ namespace Database.Repositories
 
         public async Task DeletePrescriptionByIdAsync(Guid id)
         {
-            var Prescription = new PrescriptionModel { Id = id };
+            var Prescription = GetPrescriptionById(id);
+            if (Prescription == null)
+            {
+                return;
+            }
             _context.Prescriptions.Remove(Prescription);
             await _context.SaveChangesAsync();
         }
